# language/proto: move known_imports from compiletime to runtime