… internal/go_repository_tools_srcs.bzl
… language/proto/gen/def.bzl
… language/proto/gen/gen_known_imports.go
… language/proto/known_go_imports.go
… language/proto/known_imports.go
… language/proto/known_proto_imports.go
… language/proto/proto_csv.go