**What type of PR is this?**

> Bug fix / Improvement

**What package or component does this PR mostly affect?**

> language/proto

**Which issues(s) does this PR fix?**

Fixes #1306 

**What does this PR do? Why is it needed?**

This moves the known_imports feature from compiletime to runtime.  Rather than generating go source code from the proto.csv file, it embeds the csv file into the binary and parses it in a lazy manner.

### Procedure: 

1. `bazel build //cmd/gazelle`
2.  make a trivial change in the `language/proto` package
3. `bazel build //cmd/gazelle` again.

### Before (master):

```
$ bazel build //cmd/gazelle
Target //cmd/gazelle:gazelle up-to-date:
  bazel-bin/cmd/gazelle/gazelle_/gazelle
INFO: Elapsed time: 17.915s, Critical Path: 17.28s
INFO: 5 processes: 1 internal, 4 darwin-sandbox.
INFO: Build completed successfully, 5 total actions
```

```
$ du -sh bazel-bin/cmd/gazelle/gazelle_/gazelle 78M	bazel-bin/cmd/gazelle/gazelle_/gazelle
```

### After (this PR):

```
$ bazel build //cmd/gazelle
Target //:gazelle up-to-date:
  bazel-bin/gazelle-runner.bash
  bazel-bin/gazelle
INFO: Elapsed time: 2.458s, Critical Path: 1.82s
INFO: 5 processes: 1 internal, 4 darwin-sandbox.
INFO: Build completed successfully, 5 total actions
```

```
$ du -sh bazel-bin/cmd/gazelle/gazelle_/gazelle 10M	bazel-bin/cmd/gazelle/gazelle_/gazelle
```

## Benchmark

```
Running tool: /opt/homebrew/opt/go/libexec/bin/go test -benchmem -run=^$ -bench ^BenchmarkParseCSV$ github.com/bazelbuild/bazel-gazelle/language/proto

goos: darwin
goarch: arm64
pkg: github.com/bazelbuild/bazel-gazelle/language/proto
BenchmarkParseCSV-10    	     105	  11253252 ns/op	 2454368 B/op	   12792 allocs/op
PASS
ok  	github.com/bazelbuild/bazel-gazelle/language/proto	2.157s
```

So, with this PR there is a **15s** decrease in compile time, **68MB** decrease in compiled binary size in exchange for **11ms** increase at runtime (which might be avoided depending on user configuration).

I did not profile the memory requirements during bazel build compilation, but I suspect it is significantly lower.