@achew22 2022-07-20
Maybe I'm mistaken, but I believe this means that you can no longer `go install` gazelle. You need to have a working Bazel in order to bootstrap your way in. Also, out of curiosity, does this mean you have to distribute the binary with a CSV describing the known imports?