@achew22 2022-07-20
Oh, I'm also curious, how does this impact performance? Is there any perceptible difference in startup time for Gazelle since it has to parse this?