@pcj 2022-07-20
> but I believe this means that you can no longer go install gazelle

No, it's still go installable.  My initial attempt used `go:embed` to embed the csv file, but it turned out that was incompatible with the method used by `internal/go_repository_tools.bzl` (for the `go_repository` rule).  The sources are symlinked prior to the compile step, but `go:embed` refuses any files across a symlink boundary.  As such, there is a new file `proto_csv.go` which simply contains the CSV string.

> how does this impact performance? Is there any perceptible difference in startup time

I can't seem to tell a difference.  From the benchmark test, it takes 11-12ms to parse the CSV.  This is done lazily, so startup time is not affected.  The first time a call to `knownImport` will incur the 11ms penalty.