@jfirebaugh 2022-09-07
Anything we can help with to get this PR merged?

We've been running a patched version of Gazelle at Figma with this change to avoid it getting OOM killed. Seems to be working.