@achew22 2022-09-19
I've just merged https://github.com/bazelbuild/bazel-gazelle/pull/1333. Thanks for all your hard work on this one @pcj. It was great to see people engaging with the problems they're having and sending in fixes.